Change member route to a rename_member route
since we now need to keep track of ClubMember.total_earnings, it no
longer makes sense to allow directly setting a member's balance.
We now have routes for debiting and crediting, so this route no longer
needs to be able to set the balance. Its only remaining functionality is
to rename the member